listing label pre-caching finished

+++ client/src/components/Labels/Label.tsx
@@ -4,18 +4,38 @@ import "./Label.less";
 
 type LabelProps = {
   isEditing: boolean;
-  deleteLabel: (id: string) => void;
   label: Label;
+  faint?: boolean;
+  deleteLabel?: (id: string) => void;
+  onClick?: () => void;
   className?: string;
 };
 
-const Label: React.FC<LabelProps> = ({ isEditing, deleteLabel, label }) => {
+const Label: React.FC<LabelProps> = ({
+  isEditing,
+  faint,
+  deleteLabel,
+  label,
+  onClick,
+}) => {
   const handleDeleteClick = () => {
-    deleteLabel(label.id);
+    if (deleteLabel) {
+      deleteLabel(label.id);
+    }
+  };
+  const handleClick = () => {
+    if (isEditing && onClick) {
+      onClick();
+    }
   };
 
   return (
-    <div className={`label f-center ${isEditing ? "show-delete" : ""}`}>
+    <div
+      onClick={handleClick}
+      className={`label f-center ${
+        isEditing && deleteLabel ? "show-delete" : ""
+      } ${faint && "faint"}`}
+    >
       <span>{label.name}</span>
       <button className="delete-label" onClick={handleDeleteClick}>
         x
